API: Disable CORS in production and allow everything during dev
Currently the API has no authentication and comes bundled with the GUI
in production (i.e. they will be served from the same domain, probably
localhost:8080). Therefore there is no need for CORS.

In the future this could be revisited by having API configuration
options to set the allowed origins etc.

The reason for enabling CORS without restrictions during development is
that the GUI is usually started on localhost:3000, but could run on a
different port or even a different domain, such as when using Gitpod.

Issue #26