Gitpod: Set visibility of frontend port to 'public' as well (#27)
In theory it could stay as private, but that would just give a false
sense of security/privacy, when in fact the API is kind of accessible
(of course only if you know the random URL of the Gitpod workspace).
Anyway, we are talking about Gitpod development environments, not
productive deployments.