Frontend: Enable TypeScript for the Jest test runner (#10)
https://jestjs.io/docs/getting-started#using-typescript

Instead of Babel I chose to use ts-jest to ensure the tests get type
checked while they are run.

    npm install --save-dev ts-jest
    npx ts-jest config:init

Instead of jest/globals I chose to use the types from DefinitelyTyped
because they don't require an explicit import. Apparently there is a
risk of having the versions not match anymore, but I hope to minimize
that by updating everything frequently.

    npm install --save-dev @types/jest