cleosrv: Replace native SQLite with Go-only glebarez/sqlite (#100)
Under the hood it uses https://gitlab.com/cznic/sqlite , which is a pure
Go SQLite version.

The reason for switching is that building cross-platform packages is a
real pain and I can't deal with it right now. I would prefer sticking
with the _proper_ SQLite and making the build work somehow, in which
case I would revert this. Check #100 for updates on that.