Frontend: Make 'Footer' test less flaky and more idiomatic
Note that the expect().toBeInDocument() is technically redundant because
getByText() throws an exception, causing the test to fail, if it can't
find the element.

It is still recommended to do this anyway to more clearly document the
test expectation. See
https://kentcdodds.com/blog/common-mistakes-with-react-testing-library#using-get-variants-as-assertions

The 'contentinfo' role is the default for <footer> and contains
identifying information such as Copyright notice etc. so it seems
appropriate. See:
https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/contentinfo_role