Docker: Use and create the DB in a volume and more (#110)
This is to ensure it is persisted even when re-starting or updating a
container.

Also store a config file in the volume so it can be modified, but note
that currently there are no options that are not already being specified
via CLI params. Also the default address localhost:8080 is overridden to
0.0.0.0:8080 within Docker.

Add CMD to allow passing custom parameters to cleosrv.

Document dangers with Docker anonymous volumes