build: replace make (Makefile) with mage (magefile.go) (#42)
Magefiles are better than Makefiles because they are written in Go, with
all the flexibility that gives you, and they are cross platform.

For GitHub spaces replace postCreateCommand with updateContentCommand
because it seems more appropriate ("It executes inside the container
after onCreateCommand whenever new content is available in the source
tree during the creation process.").