e2e: make the tests much more configurable
The Cypress tests now take the baseUrl and cleocPath as parameters.

Add mage targets that allow specifying which browser to use (instead of
defaulting to firefox), whether to test against some URL (instead of
starting cleosrv as part of the test) and whether to use the already
built files or new ones.

It should now also be possible to run the tests on other platforms e.g.
Windows because the paths are no longer hardcoded, but this has not been
tested.