refactor(cleosrv): use Ent to define DB model and generate schema
    go run -mod=mod entgo.io/ent/cmd/ent init Forecast \
        Estimate Probability Outcome

In most cases the files being actively edited are only
cleosrv/ent/schema/* with much of the rest being generated by Ent.

schema.graphql is now also a generated file! Manual changes to the
GraphQL schema are placed in extended.graphql .

Explicitly use IntID type for IDs as done in the ent example.

The reason for using ent (thereby replacing GORM) is:

* Includes GraphQL support, including pagination, creating mutations
  etc.
* Includes a 'Privacy' layer that could later be used to implement
  permissions
* Modelling the DB structure as a graph with strong typing and some
  included validation seems nice
* Good built-in support for DB migrations

On the whole the move to ent could be considered an experiment and if
necessary it can be undone. It was more cumbersome than expected to be
honest. Migrating the actual SQLite database schema from Cleodora
version 0.1.1 to what will become 0.2.0 will be necessary!