test: always use UTC time stamps to avoid errors
On GitHub some tests are failing:

    "{\"errors\":[{\"message\":\"sql: Scan error on column index 3, name
    \\\"created\\\": unsupported Scan, storing driver.Value type string
    into type
    *time.Time\",\"path\":[\"forecasts\"]}],\"data\":{\"forecasts\":[]}}"
    does not contain "Fabelmans"

I assume this is because the time stamp is being stored with some weird
time zone info in the DB and then can't be read back.

time.Parse() does some automatic conversions based on the current local
time zone, which would explain why it works at home but not on GitHub
(which presumably is running in a different time zone).