revert: replace GORM with ent and use it to generate GraphQL (#231)
This reverts the following commits (newest at the top):

    d96bc55 test: always use UTC time stamps to avoid errors
    db51556 docs: document changes due to ent (replacing GORM)
    e4275a9 refactor(cleosrv): remove obsolete model.go
    9c47f01 feat(cleosrv): allow Estimate.Reason to be left empty
    2450379 refactor(frontend): add support for new ent-based GraphQL
        schema
    814e7ab refactor(cleoc): make it compatible with new ent
    248d96a refactor(cleosrv): use Ent to define DB model and generate
        schema
    2d99ce4 build: ensure cleosrv code is generated first to update
        GraphQL schema

The reason for reverting is that working with ent and GraphQL is not as
nice as I initially thought so I will move all of this to a separate
branch. On 'master' it will be reverted for now. It may end up going
back into 'master'.