test: use cache=shared for test DBs to make transactions work
It's not entirely clear to me whether it is a problem that multiple
tests use the same in-memory DB (e.g. do they all run in the same
process? is it re-created correctly because the DB object is
re-created?) so giving them unique test-specific names seems to be the
safest bet.

Without this, when starting a new transaction a new DB connection is
opened, which leads to a new instance of the in-memory DB.

The error seen then is:

    no such table: forecasts