test(integration): refactor to use new client and delete obsolete tests
Some tests could not easily be replicated with the new GraphQL client
because the inputs are invalid according to the schema. This is also the
reason for removing these tests. They were just verifying that the
schema is being enforced, which is outside the scope of this project but
rather a responsibility of gqlgen.