fix(cleosrv): don't allow NewForecast.created and resolves to be zero
This was causing issues not when storing the data, but when retrieving
it back. gqlgen interprets the zero value of time.Time{} as null when
reading back from the DB, which is not allowed for these values
according to the schema.

In general it does not make sense for these values to be the zero value.

Also make some other minor related adjustements such as ensuring that
Forecast.closes is stored as nil in the DB instead of with the zero
value because then it's easier to work with the DB itself, in case that
is ever necessary.