fix(frontend): last outcome field was not being focused after delete
When deleting and adding outcomes to the form the last one was not
always being focused. It's somehow related to the fact that the length
of the 'ref' list matches the maximum old value. e.g. if at some point
there were 3 outcomes it will try to always set focus on the field with
index 4 even if now there are only 2 outcomes.