fix(frontend): ForecastDetails close button on mobile (#256)
The button was not working on Android. It's not clear to me why this
fixes it, but it turns out that the Dialog onClose() function has a
different signature than the Button onClose() and this somehow fixes it
(even though the fix does not change the Button).

This answer helped: https://stackoverflow.com/a/69756286