tests(frontend): fix 'Cannot log after tests are done' (#290)
Contrary to what I initially believed the issue does not seem to be in
the Footer.test.tsx but in App.test.tsx . The error message appears
after the Footer.test.tsx output and before the App.test.tsx output. The
App test includes all components but was not checking for the footer
content. My hypothesis is that when the test was done it tore down the
msw handlers making the fetch request for config.json made in the Footer
either abort or not be intercepted by msw. What set me on the right
track was the following issue, in particular this comment
https://github.com/jestjs/jest/issues/9324#issuecomment-1199457498 .

Before this fix the tests would fail about once every 10 to 40
executions. After this fix the tests failed due to #561 after about 320
and after about 60 executions. There were no more failures because of
this issue.