tests(frontend): fix flaky ResolveForecastDialog test (#561)
The handleClose function is only called once the resolveForecast request
completes so there was a race condition where the handleClose was not
always called immediately after clicking 'Save' in the test.

Solve this by using waitFor() to repeat the verification several times.

Also replace the previous handleClose function with the
handleCloseCalled variable with the more elegant Jest mock functions.

To simulate real network conditions a small delay was left in the
response code.