refactor: add errors package that wraps cockroachdb/errors (#112)
cockroachdb/errors gives us stack traces and information for end users
with WithHint().

The reason for the 'errors' wrapper package is so as not to spread a
third party dependency all over the code and make it easier to replace
should that ever become necessary. If the API of the internal 'errors'
keeps growing and becoming closer and closer to cockroachdb/errors then
at some point the internal package might be deleted.

In schema.resolvers.go it was necessary to import the packages as
'errors2' because otherwise the automated code generation kept
overriding it.